Reverting spine files

+++ cocos/editor-support/spine/spine-cocos2dx.cpp
@@ -48,7 +48,7 @@ void _spAtlasPage_disposeTexture (spAtlasPage* self) {
 char* _spUtil_readFile (const char* path, int* length) {
 	Data data = FileUtils::getInstance()->getDataFromFile(
 			FileUtils::getInstance()->fullPathForFilename(path).c_str());
-	*length = (int) data.getSize();
+	*length = data.getSize();
 	char* bytes = MALLOC(char, *length);
 	memcpy(bytes, data.getBytes(), *length);
 	return bytes;
